backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
huggingface-hub==0.23.4
pydantic==2.7.4
python-dotenv==1.0.1
slowapi==0.1.9
